Hack it working...

+++ isc-dhcp-script.c
@@ -18,7 +18,7 @@
 #include <linux/rtnetlink.h>
 
 struct cb_data {
-    int index;
+    unsigned int index;
     bool permanent;
 };
 
@@ -122,7 +122,7 @@ main(int argc, char **argv)
     char key[128];
     for (;*cur; cur++) {
         char *end = strchr(*cur, '=');
-        if (!end || end - *cur >= sizeof(key))
+        if (!end || end - *cur >= (int) sizeof(key))
             continue; // Ignore environment entries with very long or invalid keys
         memcpy(key, *cur, end - *cur);
         key[end - *cur] = 0;
